[FIX] estate: fix spacing issues and remove unused code
This commit addresses minor spacing inconsistencies and removes
unused code fragments in the estate module. These changes improve
readability and maintain code cleanliness.

+++ estate/models/estate_property.py
@@ -34,7 +34,6 @@ class EstateProperty(models.Model):
             ('offer_accepted', 'Offer Accepted'),
             ('sold', 'Sold'),
             ('cancelled', 'Cancelled')],
-
             required=True,
             copy=False,
             default='new'
@@ -46,7 +45,7 @@ class EstateProperty(models.Model):
     salesperson_id = fields.Many2one('res.users', string='Salesman', tracking=True, default=lambda self: self.env.user)
     company_id = fields.Many2one("res.company", default=lambda self: self.env.company, required=True)
     buyer_id = fields.Many2one('res.partner', string='Buyer', copy=False, tracking=True)
-    total_area = fields.Float(compute='_compute_total_area', string='Total Area (sqm)')
+    total_area = fields.Integer(compute='_compute_total_area', string='Total Area (sqm)')
     best_price = fields.Float(compute='_compute_best_offer', string='Best Price')
 
     _sql_constraints = [
